KVM: selftests: Test unsigned timer comparison when CNT > CVAL
Test that hardware implements the timer condition of the Generic Timer
as an unsigned comparison by offsetting CNTVCT to a 'negative' value if
interpreted as a signed integer. The expected comparison:

  TimerConditionMet = (CNTVCT_EL0 >= CNTV_CVAL_EL0)

which matches the KVM implementation. Note that this diverges from the
Arm ARM's somewhat vague description of the timer condition, which says:

  TimerConditionMet = (((Counter[63:0] - Offset[63:0]) - CompareValue[63:0]) >= 0)

implying that the difference between CNTVCT and CVAL is a signed
quantity. This test passes on the Arm fast model configured for 8.6:

  ./FVP_Base_RevC-2xAEMvA \
	-C pctl.startup=0.*.*.* \
	-C cache_state_modelled=0 \
	-C bp.refcounter.non_arch_start_at_default=1 \
	-C bp.secure_memory=false \
	-C cluster0.has_arm_v8-6=1 \
	-C cluster1.has_arm_v8-6=1

as well as the M1 Ultra and Neoverse N1. It fails on an upcoming 8.6
implementation. All credit to the folks who spotted this issue and root
caused it.

Reported-by: Colton Lewis <coltonlewis@google.com>
Reported-by: Ricardo Koller <ricarkol@google.com>
Signed-off-by: Oliver Upton <oliver.upton@linux.dev>
Change-Id: Iad5626526175843b0c4a326245c596066af4ca2c