KVM: arm64: Only return attributes from stage2_update_leaf_attrs()
Returning a single PTE from stage2_update_leaf_attrs() doesn't make a
great deal of sense given that the function could be used to apply a
change to a range of PTEs. Instead, return a bitwise OR of attributes
from all the visited PTEs.

As the walker is no longer returning the full PTE, drop the check for a
valid PTE in kvm_age_gfn().

Signed-off-by: Oliver Upton <oliver.upton@linux.dev>