arm64: cpu_errata: Mitigate Ampere erratum PILE_OF_SAND
Certain Ampere designs have a broken HAFDBS implementation where
hardware updates to a descriptor is done as a non-atomic access instead
of a compare-exchange on the entire descriptor. Of course, this can race
with software updates to the descriptor, which could result in hardware
erroneously setting bit 10 on a table or invalid descriptor. Affected
designs do _not_ report support for HAFDBS in MMFR1, however the
underlying system register controls are still functional.

Not only does this conflict with the architecture, it mixes poorly with
KVM's aggressive setting of VTCR_EL2.HA, which relies on the bit being
RES0 on systems that don't support the feature.

Add a workaround for the relevant Ampere parts, avoiding the rather
catastrophic effects that can result from the broken HAFDBS
implementation. Additionally, expand the comment on VTCR_EL2.HA and
explain why KVM sets the bit unconditionally.

Cc: stable@vger.kernel.org
Signed-off-by: Oliver Upton <oliver.upton@linux.dev>